flask_server/requirements_yolov5.txt
ultralytics
opencv-python
paho-mqtt
numpy
flask